refactor(bot-client): address third PR review round
- Narrow joinFooter type from (string|number|boolean|null|undefined)
  to (string|false|null|undefined) — compiler now catches the
  count && pluralize() footgun instead of silently swallowing 0
- Extract footer tests to colocated footer.test.ts per code standards
- Fix deny/browse footer: count+filter are one segment (no extra •
  delimiter between them), matching original UX
